refactored code to extract a specified portion of a wav file
